Improves build-mmr execution in different environments
Allows the `build-mmr` command to be executed in ECS
environments where environment variables might not be set
via a file. It now checks for an `ENV_FILE`, then a default
docker env file, and finally falls back to direct environment
variable usage.

Additionally, this commit disables the `ipfs` service in the
docker-compose file.